+++ src/graph/graph_runner.ts
@@ -16,10 +16,11 @@
  */
 
 import {InputProvider} from '../data/input_provider';
-import {tidy} from '../math/backends/tracking';
+import {tidy} from '../globals';
 import {NDArrayMath} from '../math/math';
-import {Tensor, Scalar} from '../math/tensor';
 import {Optimizer} from '../math/optimizers/optimizer';
+import {Scalar, Tensor} from '../math/tensor';
+
 import {SymbolicTensor} from './graph';
 import {CostReduction, FeedEntry, Session} from './session';
