+++ src/graph/ops/convolution.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {keep, tidy} from '../../math/backends/tracking';
+import {keep, tidy} from '../../globals';
 import * as conv_util from '../../math/conv_util';
 import {NDArrayMath} from '../../math/math';
 import {Tensor1D, Tensor3D, Tensor4D} from '../../math/tensor';
