+++ src/graph/ops/linear_combination.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {keep, tidy} from '../../math/backends/tracking';
+import {keep, tidy} from '../../globals';
 import {NDArrayMath} from '../../math/math';
 import {Scalar} from '../../math/tensor';
 import {SymbolicTensor} from '../graph';
