+++ src/graph/ops/log.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {keep, tidy} from '../../math/backends/tracking';
+import {keep, tidy} from '../../globals';
 import {NDArrayMath} from '../../math/math';
 import {SymbolicTensor} from '../graph';
 import * as graph_util from '../graph_util';
