+++ src/graph/ops/reshape.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {keep, tidy} from '../../math/backends/tracking';
+import {keep, tidy} from '../../globals';
 import {NDArrayMath} from '../../math/math';
 import {Tensor} from '../../math/tensor';
 import * as util from '../../util';
