+++ src/graph/optimizers/adam_optimizer.ts
@@ -15,10 +15,10 @@
  * =============================================================================
  */
 
-import {keep, tidy} from '../../math/backends/tracking';
+import {keep, tidy} from '../../globals';
 import {NDArrayMath} from '../../math/math';
-import {Tensor, Scalar} from '../../math/tensor';
 import {Optimizer} from '../../math/optimizers/optimizer';
+import {Scalar, Tensor} from '../../math/tensor';
 import {NamedVariableMap} from '../../math/types';
 import {Node} from '../graph';
 import {SessionRuntime} from '../session';
