+++ src/math/backends/tracking.ts
@@ -15,8 +15,86 @@
  * =============================================================================
  */
 
-import * as tracking_functions from './tracking_functions'
+import {ENV} from '../../environment';
+import {doc} from '../decorators';
+import {Tensor} from '../tensor';
 
-export const tidy = tracking_functions.Tracking.tidy;
-export const keep = tracking_functions.Tracking.keep;
-export const time = tracking_functions.Tracking.time;
+import {ScopeFn, ScopeResult, ScopeResultImmediate} from './tape_util';
+
+export class Tracking {
+  /**
+   * Executes the provided function and after it is executed, cleans up all
+   * intermediate tensors allocated by the function except those returned by
+   * the function.
+   *
+   * Using this method helps avoid memory leaks. In general, wrap calls to
+   * operations in dl.tidy() for automatic memory cleanup.
+   *
+   * When in safe mode, you must enclose all `Tensor` creation and ops
+   * inside a `dl.tidy()` to prevent memory leaks.
+   *
+   * @param nameOrFn The name of the closure, or the function to execute.
+   *     If a name is provided, the 2nd argument should be the function.
+   *     If a name is provided, and debug mode is on, the timing and the memory
+   *     usage of the function will be tracked and displayed on the console
+   *     using the provided name.
+   * @param fn The function to execute.
+   * @param gradientsMode If true, enables gradients mode.
+   *     See dl.gradientsScope for details.
+   */
+  @doc({heading: 'Performance', subheading: 'Memory'})
+  static tidy<T extends ScopeResult>(
+      nameOrFn: string|ScopeFn<T>, fn?: ScopeFn<T>, gradientsMode = false): T {
+    if (fn == null) {
+      // Called with only 1 argument.
+      if (typeof nameOrFn !== 'function') {
+        throw new Error('Please provide a function to dl.tidy()');
+      }
+      fn = nameOrFn;
+      nameOrFn = '';
+    } else {
+      // Called with 2 arguments.
+      if (typeof nameOrFn !== 'string' && !(nameOrFn instanceof String)) {
+        throw new Error(
+            'When calling with two arguments, the first argument ' +
+            'to dl.tidy() must be a string');
+      }
+      if (typeof fn !== 'function') {
+        throw new Error(
+            'When calling with two arguments, the 2nd argument ' +
+            'to dl.tidy() must be a function');
+      }
+      // TODO(nsthorat,smilkov): Do operation logging and performance profiling.
+    }
+    ENV.engine.startScope(gradientsMode);
+
+    const result = fn();
+    if (result instanceof Promise) {
+      result.then(r => ENV.engine.endScope(r, gradientsMode));
+      return result;
+    } else {
+      ENV.engine.endScope(result as ScopeResultImmediate, gradientsMode);
+      return result;
+    }
+  }
+
+  /**
+   * Keeps a Tensor generated inside a dl.tidy() from being disposed
+   * automatically.
+   * @param result The Tensor to keep from being disposed.
+   */
+  @doc({heading: 'Performance', subheading: 'Memory'})
+  static keep<T extends Tensor>(result: T): T {
+    return ENV.engine.keep(result);
+  }
+
+  /**
+   * Executes f() and returns a promise that resolves with the elapsed time of
+   * f() in milliseconds.
+   * @param f The function to execute and time.
+   */
+  @doc({heading: 'Performance', subheading: 'Timing'})
+  static time(f: () => void): Promise<number> {
+    return ENV.backend.time(f);
+  }
+}
